auth reformulation
session removed -> userToken added
auth removed from SessionBusiness to AuthenticationProvider
