+++ APP/SimpleBank.AcctManage/SimpleBank.AcctManage.API/Config/Services/ConfigServicePersistence.cs
@@ -19,7 +19,7 @@ namespace SimpleBank.AcctManage.API.Config.Services
             services.AddScoped<ITransferRepository, TransferRepository>();
             services.AddScoped<IMovementRepository, MovementRepository>();
             services.AddScoped<IAccountRepository, AccountRepository>();
-            services.AddScoped<ISessionRepository, SessionRepository>();
+            services.AddScoped<IUserTokenRepository, UserTokenRepository>();
 
             services.AddScoped<IUnitOfWork, UnitOfWork>();
