+++ APP/SimpleBank.AcctManage/SimpleBank.AcctManage.API/DTModels/Requests/LogoutUserRequest.cs
@@ -1,9 +1,8 @@
 ﻿namespace SimpleBank.AcctManage.API.DTModels.Requests
 {
-
     public class LogoutUserRequest
     {
-        public Guid SessionId { get; set; }
+        public Guid UserTokenId { get; set; }
 
     }
 }
