--- APP/SimpleBank.AcctManage/SimpleBank.AcctManage.API/_Assets/SimpleBankAPI_pt3.txt
@@ -1,91 +1,0 @@
-
--- Database: SimpleBankAPI
-
--- DROP DATABASE IF EXISTS "SimpleBankAPI";
-/*
-CREATE DATABASE "SimpleBankAPI"
-    WITH
-    OWNER = postgres
-    ENCODING = 'UTF8'
-    LC_COLLATE = 'English_United Kingdom.1252'
-    LC_CTYPE = 'English_United Kingdom.1252'
-    TABLESPACE = pg_default
-    CONNECTION LIMIT = -1;
-
-COMMENT ON DATABASE "SimpleBankAPI"
-    IS 'Simple Bank API';
-    
-*/
-
-CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
-
-DROP TABLE IF EXISTS "Transfers";
-DROP TABLE IF EXISTS "Movements";
-DROP TABLE IF EXISTS "Accounts";
-DROP TABLE IF EXISTS "Sessions";
-DROP TABLE IF EXISTS "Users";
-
-CREATE TABLE "Users"(
-   User_id       uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
-   Email            VARCHAR(320) NOT NULL,
-   Fullname                 TEXT NOT NULL,
-   "Password"              BYTEA NOT NULL,
-   Salt                    BYTEA NOT NULL,
-   Username                 TEXT NOT NULL,
-   Created_at          TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
-   Password_changed_at TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
-CONSTRAINT Check_MinLength_Users_FullName CHECK (LENGTH(FullName) >= 16),
-CONSTRAINT Check_MinLength_Users_Username CHECK (LENGTH(Username) >= 8),
-CONSTRAINT UNIQUE_Users_Email    UNIQUE (Email),
-CONSTRAINT UNIQUE_Users_Username UNIQUE (Username)
-);
-    
-CREATE TABLE "Accounts"(
-   Account_id uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
-   User_id               uuid NOT NULL,
-   Balance            NUMERIC NOT NULL,
-   Created_at       TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
-   Currency           CHAR(3) NOT NULL DEFAULT ('EUR'),
-CONSTRAINT Check_Accounts_Balance_Positive CHECK (Balance > 0),
-CONSTRAINT Account_User_fkey FOREIGN KEY (User_id)
-    REFERENCES "Users" (User_id) MATCH SIMPLE
-    ON UPDATE NO ACTION ON DELETE NO ACTION,
-CONSTRAINT Check_Accounts_Currency_Length CHECK (LENGTH(Currency) = 3)
-);
-    
-CREATE TABLE "Movements"(
-   Movement_id    uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
-   Account_id                uuid NOT NULL,
-   Amount                 NUMERIC NOT NULL,
-   Created_at           TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
-CONSTRAINT Movements_Accounts_fkey FOREIGN KEY (Account_id)
-    REFERENCES "Accounts" (Account_id) MATCH SIMPLE
-    ON UPDATE NO ACTION ON DELETE NO ACTION
-);  
-    
-CREATE TABLE "Transfers"(
-    Transfer_id    uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
-    Amount                 NUMERIC NOT NULL,
-    From_account_id           uuid NOT NULL,
-    To_account_id             uuid NOT NULL,
-    Created_at         TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
-CONSTRAINT Check_Transfers_Amount_Positive            CHECK (Amount > 0),
-CONSTRAINT Check_Transfers_FromAndToAccounts_NotEqual CHECK (From_account_id <> To_account_id)
-);  
-
-CREATE TABLE "Sessions"(
-   Session_id         uuid PRIMARY KEY DEFAULT uuid_generate_v4 (), --
-   User_id                       uuid NOT NULL, 
-   Active                        BOOL NOT NULL,
-   Created_at               TIMESTAMP NOT NULL,
-   Access_token                  TEXT NOT NULL,
-   Access_token_expires_at  TIMESTAMP NOT NULL,
-   Refresh_token                 TEXT NOT NULL,
-   Refresh_token_expires_at TIMESTAMP NOT NULL,
-CONSTRAINT Sessions_Users_fkey FOREIGN KEY (User_id)
-    REFERENCES "Users" (User_id) MATCH SIMPLE
-    ON UPDATE NO ACTION ON DELETE NO ACTION
-);
-
-    
-
