+++ APP/SimpleBank.AcctManage/SimpleBank.AcctManage.API/appsettings.json
@@ -13,7 +13,7 @@
     "SecretKey": "IamAsecretIfeelSoSpecialAboutMyselfIamSupposedToBeAlittleBitLongerMaybeHaverAnumber32Yeeh",
     "Issuer": "https://localhost:7025",
     "Audience": "simplebankapi",
-    "AccessDuration": 119,
+    "AccessDuration": 60,
     "RefreshDuration": 120
   },
   "KafkaTopic": "transfer",
