--- APP/SimpleBank.AcctManage/SimpleBank.AcctManage.Core.Application/Contracts/Persistence/ISessionRepository.cs
@@ -1,11 +1,0 @@
-﻿using SimpleBank.AcctManage.Core.Domain;
-
-namespace SimpleBank.AcctManage.Core.Application.Contracts.Persistence
-{
-    public interface ISessionRepository : ICommonRepository<Session>
-    {
-        Task<Session?> GetLastSessionOfUserAsync(Guid userId);
-        Session? GetLastSessionOfUser(Guid userId);
-
-    }
-}
