+++ APP/SimpleBank.AcctManage/SimpleBank.AcctManage.Core.Application/Contracts/Providers/IAuthenthicationProvider.cs
@@ -1,0 +1,12 @@
+﻿using SimpleBank.AcctManage.Core.Domain;
+
+namespace SimpleBank.AcctManage.Core.Application.Contracts.Providers
+{
+    public interface IAuthenthicationProvider
+    {
+        Task<(UserToken?, string?)> ProcessLoginAsync(Guid userId);
+        Task<(UserToken?, string?)> ProcessRenewToken(string refreshToken);
+        Task<bool?> ProcessLogout(Guid userTokenId);
+
+    }
+}
