+++ APP/SimpleBank.AcctManage/SimpleBank.AcctManage.Infrastructure.Persistence/Configurations/AccountConfiguration.cs
@@ -8,9 +8,12 @@ namespace SimpleBank.AcctManage.Infrastructure.Persistence.Configurations
     {
         public void Configure(EntityTypeBuilder<Account> builder)
         {
-            builder.Property(e => e.Id).HasColumnName("account_id");
+            builder.Property(e => e.Id)
+                .HasColumnName("account_id")
+                .HasDefaultValueSql("uuid_generate_v4()"); 
 
-            builder.Property(e => e.Balance).HasColumnName("balance");
+            builder.Property(e => e.Balance)
+                .HasColumnName("balance");
 
             builder.Property(e => e.CreatedAt)
                 .HasColumnType("timestamp without time zone")
@@ -23,7 +26,8 @@ namespace SimpleBank.AcctManage.Infrastructure.Persistence.Configurations
                 .HasDefaultValueSql("'EUR'::bpchar")
                 .IsFixedLength();
 
-            builder.Property(e => e.UserId).HasColumnName("user_id");
+            builder.Property(e => e.UserId)
+                .HasColumnName("user_id");
 
             builder.HasOne(d => d.User)
                 .WithMany(p => p.Accounts)
