+++ APP/SimpleBank.AcctManage/SimpleBank.AcctManage.Infrastructure.Persistence/Configurations/MovementConfiguration.cs
@@ -8,11 +8,15 @@ namespace SimpleBank.AcctManage.Infrastructure.Persistence.Configurations
     {
         public void Configure(EntityTypeBuilder<Movement> builder)
         {
-            builder.Property(e => e.Id).HasColumnName("movement_id");
+            builder.Property(e => e.Id)
+                .HasColumnName("movement_id")
+                .HasDefaultValueSql("uuid_generate_v4()"); 
 
-            builder.Property(e => e.AccountId).HasColumnName("account_id");
+            builder.Property(e => e.AccountId)
+                .HasColumnName("account_id");
 
-            builder.Property(e => e.Amount).HasColumnName("amount");
+            builder.Property(e => e.Amount)
+                .HasColumnName("amount");
 
             builder.Property(e => e.CreatedAt)
                 .HasColumnType("timestamp without time zone")
