+++ APP/SimpleBank.AcctManage/SimpleBank.AcctManage.Infrastructure.Persistence/Repositories/UserTokenRepository.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using SimpleBank.AcctManage.Core.Domain;
+﻿using SimpleBank.AcctManage.Core.Domain;
 using SimpleBank.AcctManage.Core.Application.Contracts.Persistence;
 using SimpleBank.AcctManage.Infrastructure.Persistence.Repositories.Common;
 using Serilog;
