+++ APP/SimpleBank.AcctManage/_init_project/docker_postgres_init.sql
@@ -6,13 +6,13 @@ CREATE SCHEMA "SB-operational" AUTHORIZATION postgres;
 COMMENT ON SCHEMA "SB-operational" IS 'All data related to account management operations.';
 
 CREATE TABLE "SB-operational"."Users"(
-	User_id       uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
-	Created_at          TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
-	Username                 TEXT NOT NULL,
-	Email            VARCHAR(320) NOT NULL,
-	Fullname                 TEXT NOT NULL,
-	"Password"              BYTEA NOT NULL,
-	Salt                    BYTEA NOT NULL,
+	User_id            uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
+	Created_at            TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
+	Username                   TEXT NOT NULL,
+	Email              VARCHAR(320) NOT NULL,
+	Fullname                   TEXT NOT NULL,
+	"Password"                BYTEA NOT NULL,
+	Salt                      BYTEA NOT NULL,
 	Password_changed_at TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
 CONSTRAINT Check_MinLength_Users_FullName CHECK (LENGTH(FullName) >= 16),
 CONSTRAINT Check_MinLength_Users_Username CHECK (LENGTH(Username) >= 8),
@@ -34,7 +34,7 @@ CONSTRAINT Check_Accounts_Currency_Length CHECK (LENGTH(Currency) = 3)
 );
     
 CREATE TABLE "SB-operational"."Movements"(
-	Movement_id    uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
+	Movement_id       uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
 	Created_at           TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
 	Account_id                uuid NOT NULL,
 	Amount                 NUMERIC NOT NULL,
@@ -48,14 +48,14 @@ CONSTRAINT Movements_Accounts_fkey FOREIGN KEY (Account_id)
 CREATE SCHEMA "SB-auth" AUTHORIZATION postgres;
 COMMENT ON SCHEMA "SB-auth" IS 'All data related to account management authentication.';
 
-CREATE TABLE "SB-auth"."UserToken"(
-	User_token_id         uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
-	Created_at               TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
-	User_id                       uuid NOT NULL, 
-	Access_token                  TEXT NOT NULL,
-	Access_token_expires_at  TIMESTAMP NOT NULL,
-	Refresh_token                 TEXT NOT NULL,
-	Refresh_token_expires_at TIMESTAMP NOT NULL
+CREATE TABLE "SB-auth"."UserTokens"(
+	User_token_id           uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
+	Created_at                 TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
+	User_id                         uuid NOT NULL, 
+	Access_token                    TEXT NOT NULL,
+	Access_token_expires_at    TIMESTAMP NOT NULL,
+	Refresh_token                   TEXT NOT NULL,
+	Refresh_token_expires_at   TIMESTAMP NOT NULL
 );
 
 
@@ -64,7 +64,7 @@ CREATE SCHEMA "SB-historic" AUTHORIZATION postgres;
 COMMENT ON SCHEMA "SB-historic" IS 'All data related to account management history.';
 
 CREATE TABLE "SB-historic"."Transfers"(
-    Transfer_id    uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
+    Transfer_id       uuid PRIMARY KEY DEFAULT uuid_generate_v4 (),
 	Created_at           TIMESTAMP NOT NULL DEFAULT (CURRENT_TIMESTAMP),
     Amount                 NUMERIC NOT NULL,
     From_account_id           uuid NOT NULL,
